+++ packages/terra-core-docs/src/terra-dev-site/test/form-select/AllowClear.test.jsx
@@ -25,7 +25,6 @@ class AllowClear extends React.Component {
           id="allowclear"
           onChange={this.handleChange}
           placeholder="Select a color"
-          ariaLabel="Select a color"
           value={this.state.value}
           variant="combobox"
         >