+++ packages/terra-core-docs/src/terra-dev-site/test/form-select/RequiredCombobox.test.jsx
@@ -24,7 +24,6 @@ class ControlledCombobox extends React.Component {
           id="combobox"
           onChange={this.handleChange}
           placeholder="Select a color"
-          ariaLabel="Select a color"
           required
           value={this.state.value}
           variant="combobox"