+++ packages/terra-form-select/src/SearchSelect.jsx
@@ -105,10 +105,6 @@ const propTypes = {
    * The id of the input field.
    */
   inputId: PropTypes.string,
-  /**
-   * Used to define a string that labels the select component to screen readers.
-   */
-  ariaLabel: PropTypes.string,
 };
 
 const defaultProps = {
@@ -130,7 +126,6 @@ const defaultProps = {
   isTouchAccessible: false,
   value: undefined,
   inputId: undefined,
-  ariaLabel: undefined,
 };
 
 class SearchSelect extends React.Component {
@@ -207,7 +202,7 @@ class SearchSelect extends React.Component {
 
   render() {
     const {
-      allowClear, children, defaultValue, onChange, placeholder, required, value, intl, inputId, ariaLabel, ...otherProps
+      allowClear, children, defaultValue, onChange, placeholder, required, value, intl, inputId, ...otherProps
     } = this.props;
 
     const defaultPlaceholder = intl.formatMessage({ id: 'Terra.form.select.defaultDisplay' });
@@ -235,7 +230,6 @@ class SearchSelect extends React.Component {
         totalOptions={SelectUtil.getTotalNumberOfOptions(children)}
         clearOptionDisplay={clearOptionDisplay}
         inputId={inputId}
-        ariaLabel={ariaLabel}
         resetComboboxValue={this.handleResetComboboxValue}
       >
         {children}