+++ packages/terra-form-select/src/combobox/Frame.jsx
@@ -525,6 +525,7 @@ class Frame extends React.Component {
     const {
       ariaLabel,
       disabled,
+      placeholder,
       intl,
     } = this.props;
 
@@ -533,10 +534,10 @@ class Frame extends React.Component {
     // VO on iOS doesn't do a good job of announcing disabled stated. Here we append the phrase that
     // VO associates with disabled form controls.
     if ('ontouchstart' in window && disabled) {
-      return `${ariaLabel} ${dimmed}`;
+      return ariaLabel === undefined ? `${placeholder} ${dimmed}` : `${ariaLabel} ${dimmed}`;
     }
 
-    return ariaLabel;
+    return ariaLabel === undefined ? placeholder : ariaLabel;
   }
 
   /**