+++ packages/terra-core-docs/src/terra-dev-site/doc/form-select/example/combobox/Invalid.jsx
@@ -24,5 +24,5 @@ const InvalidExample = () => {
     </>
   );
 };
-export default InvalidExample;
 
+export default InvalidExample;