+++ packages/terra-form-select/CHANGELOG.md
@@ -3,8 +3,8 @@
 ## Unreleased
 
 * Added
-  * Added 'aria-invalid' attribute which will be set to true for error input fields and false when resolving errors.
-  * Added 'ariaLabel' attribute in `form-select`.
+  * Added 'aria-invalid' attribute for `terra-form-select-combobox`.
+  * Updated auto assign placeholder text to `aria-label` in `terra-form-select-combobox` and `terra-form-search-select`.
 
 ## 6.50.0 - (November 13, 2023)