+++ packages/terra-form-select/CHANGELOG.md
@@ -8,7 +8,7 @@
 ## 6.51.0 - (November 21, 2023)
 
 * Added
-  * Added 'aria-invalid' attribute which will be set to true for error input fields and false when resolving errors.
+  * Added 'aria-invalid' attribute for `terra-form-select-combobox`
 
 ## 6.50.0 - (November 13, 2023)